# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sistem Absensi Mahasiswa Geolocation API",
        "description": "API untuk sistem absensi mahasiswa berbasis geolocation. Mahasiswa dapat melakukan absensi dengan validasi lokasi GPS.",
        "contact": {
            "name": "Admin Support",
            "email": "admin@kampus.ac.id"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API Server v1"
        }
    ],
    "paths": {
...
        "/reports/attendance": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Laporan absensi",
                "description": "Mendapatkan laporan absensi dengan filter opsional (Admin only)",
                "operationId": "attendanceReport",
                "parameters": [
                    {
                        "name": "start_date",
                        "in": "query",
                        "description": "Tanggal mulai (format: Y-m-d)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2025-01-01"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "description": "Tanggal akhir (format: Y-m-d)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2025-12-31"
                        }
                    },
                    {
                        "name": "schedule_id",
                        "in": "query",
                        "description": "Filter berdasarkan jadwal",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "user": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer"
                                                            },
                                                            "name": {
                                                                "type": "string"
                                                            },
                                                            "email": {
                                                                "type": "string"
                                                            }
                                                        },
                                                        "type": "object"
                                                    },
                                                    "schedule": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer"
                                                            },
                                                            "course_name": {
                                                                "type": "string"
                                                            }
                                                        },
                                                        "type": "object"
                                                    },
                                                    "status": {
                                                        "type": "string"
                                                    },
                                                    "distance": {
                                                        "type": "number"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden - Admin only"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/reports/attendance/export": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Export laporan ke Excel",
                "description": "Download laporan absensi dalam format Excel (Admin only)",
                "operationId": "exportAttendanceExcel",
                "parameters": [
                    {
                        "name": "start_date",
                        "in": "query",
                        "description": "Tanggal mulai (format: Y-m-d)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "description": "Tanggal akhir (format: Y-m-d)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "schedule_id",
                        "in": "query",
                        "description": "Filter berdasarkan jadwal",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "File Excel berhasil didownload",
                        "content": {
                            "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet": {}
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden - Admin only"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Masukkan token yang didapat dari endpoint login",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}